// src/user/domain/entity/address/country/enum/country-language.enum.ts
enum CountryLanguageEnum {
  ARGENTINA = 'Spanish',
  BOLIVIA = 'Spanish',
  BRAZIL = 'Portuguese',
  CHILE = 'Spanish',
  COLOMBIA = 'Spanish',
  ECUADOR = 'Spanish',
  GUYANA = 'English',
  PARAGUAY = 'Spanish',
  PERU = 'Spanish',
  SURINAME = 'Dutch',
  URUGUAY = 'Spanish',
  VENEZUELA = 'Spanish',
}

export default CountryLanguageEnum;